+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   Added method to fetch resource by received event.
 *   Added log method to write to custom debuglog to heidelpay facade.
 *   Added IdService to fetch ids and types from id strings.
+*   Added Alipay payment type.
 
 ### Changed
 *   Applied several micro optimizations.